+++ src/components/App/App.js
@@ -1,12 +1,12 @@
-import Register from "./Register/Register";
-import Login from "./Login/Login";
-import Header from "./Header/Header";
-import Main from "./Main/Main";
-import Movies from "./Movies/Movies";
-import Profile from "./Profile/Profile";
-import Footer from "./Footer/Footer";
+import Register from "../Register/Register";
+import Login from "../Login/Login";
+import Header from "../Header/Header";
+import Main from "../Main/Main";
+import Movies from "../Movies/Movies";
+import Profile from "../Profile/Profile";
+import Footer from "../Footer/Footer";
 import { Route, Switch } from "react-router-dom";
-import NotFound from "./NotFound/NotFound";
+import NotFound from "../NotFound/NotFound";
 
 
 function App() {
@@ -19,34 +19,24 @@ function App() {
         <Route path="/signin">
           <Login />
         </Route>
-
         <Route exact path={['/', '/movies', '/saved-movies', '/profile']}>
           <Route exact path="/">
             <Header
-            class='header_main' />
+              class='header_main' />
             <Main />
           </Route>
           <Header />
-
           <Switch>
-
-            {/* <Route exact path="/">
-              <Main />
-            </Route> */}
-
             <Route path="/movies">
               <Movies />
             </Route>
-
             <Route path="/saved-movies">
               <Movies />
             </Route>
-
             <Route path="/profile">
               <Profile />
             </Route>
           </Switch>
-
           <Route exact path={['/', '/movies', '/saved-movies']}>
             <Footer />
           </Route>
@@ -55,9 +45,6 @@ function App() {
           <NotFound />
         </Route>
       </Switch>
-
-
-
     </div>
   );
 }
