+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter } from 'react-router-dom'; 
 import ReactDOM from 'react-dom/client';
 import "./pages/index.css";
-import App from './components/App';
+import App from './components/App/App';
 import reportWebVitals from './reportWebVitals';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
